little code restructurisation
